Add QUERY type in execute RPC
