+++ cmd/root.go
@@ -19,13 +19,6 @@ var (
 	}
 )
 
-/*
-Commands; todo
-- start / add node
-- shutdown node
-- print config
-*/
-
 func Execute() {
 	rootCmd.Flags().SortFlags = false
