+++ pkg/services/rpc/service.go
@@ -122,10 +122,10 @@ func (s service) Execute(input *domain.ExecuteInput) (*domain.ExecuteOutput, err
 	}
 
 	res := <-s.node.Execute(*input)
-	if res.Err != nil {
-		return &res, res.Err
-	}
-
+	// if res.Err != nil {
+	// 	return &res, res.Err
+	// }
+	// Should not mix error and res.Err
 	return &res, nil
 }
